move intersection component code to a separate file
